requirements.txt
asgiref==3.2.10
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.1
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
idna==3.4
Pillow==9.4.0
python-decouple==3.8
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
